Add full variant to the approximate timeout functiont
The g_timeout_add_seconds() API lacks a _full() counterpart, allowing the
setting of a destroy notification function to be invoked when the timeout
source is removed.

This patch adds g_timeout_add_seconds_full() to the public API and
reimplements g_timeout_add_seconds() as a call to g_timeout_add_seconds_full().


git-svn-id: http://svn.gnome.org/svn/glib/trunk@5575 5bbd4a9e-d125-0410-bf1d-f987e7eefc80